requirements.txt
Flask==2.0.0
Werkzeug==2.0.0
flask_restful==0.3.10
python-dotenv==1.0.1
google-generativeai==0.8.4
requests==2.32.3